Remove the Python dependency, increment the version number, and update README.md
The `python` cookbook has a dependency on the `build-essentials` and `yum`
cookbooks, so adding this dependency actually introduces several undocumented
ones. As most users of Tracelytics won't actually be installing Python, it
makes more sense to create a separate cookbook (via git branch?), or just leave
the dependency checking to the end user.